+++ taf/repository_tool.py
@@ -65,6 +65,15 @@ def is_delegated_role(role):
     return role not in ("root", "targets", "snapshot", "timestamp")
 
 
+def is_auth_repo(repo_path):
+    """Check if the given path contains a valid TUF repository"""
+    try:
+        Repository(repo_path)._repository
+        return True
+    except Exception:
+        return False
+
+
 def load_role_key(keystore, role, password=None, scheme=DEFAULT_RSA_SIGNATURE_SCHEME):
     """Loads the specified role's key from a keystore file.
     The keystore file can, but doesn't have to be password protected.
@@ -272,11 +281,15 @@ class Repository:
         """
         # before attempting to tuf repository, create empty targets directory if it does not exist
         # to avoid errors raised by tuf
+        targets_existed = True
         if not self.targets_path.is_dir():
+            targets_existed = False
             self.targets_path.mkdir(parents=True, exist_ok=True)
         try:
             self._tuf_repository = load_repository(path, self.name)
         except RepositoryError:
+            if not targets_existed:
+                self.targets_path.rmdir()
             raise InvalidRepositoryError(f"{self.name} is not a valid TUF repository!")
 
     def reload_tuf_repository(self):